% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_rmd_emoji_addin.R
\name{insert_rmd_emoji_addin}
\alias{insert_rmd_emoji_addin}
\title{Insert code snippet `r emo::ji("smile")`}
\usage{
insert_rmd_emoji_addin()
}
\description{
Call this function as an addin to insert `r emo::ji("smile")` at the cursor
position.
}
\details{
adapted from [rstudioapi::insertInAddin()]
}
